feat: place 조회 로직 구현

+++ tot/src/main/java/com/trip/place/model/mapper/PlaceMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface PlaceMapper {
 
     List<PlaceDto> selectAll();
-    PlaceDto selectById(Long id);
+    PlaceDto selectById(int id);
     void insertPlace(PlaceDto place);
     PlaceDto selectPlace(String name);
